aoi_buffer {AOI}	R Documentation

Buffer AOI

Description

Add or subtract a uniform distance to/from a spatial object in either miles or kilometers.

Usage

aoi_buffer(AOI, d, km = FALSE)

Arguments

AOI
a spatial, raster or simple features object

d
numeric.The distance by which to modify each edge

km
logical. If TRUE distances are in kilometers, default is FALSE with distances in miles

Value

a spatial geometry of the same class as the input AOI (if Raster sp returned)

Examples

## Not run: 
 # get an AOI of 'Garden of the Gods' and add a 2 mile buffer
    AOI = aoi_get("Garden of the Gods") %>% modify(2)

 # get an AOI of 'Garden of the Gods' and add a 2 kilometer buffer
    getAOI("Garden of the Gods") %>% modify(2, km = TRUE)

 # get and AOI for Colorado Springs and subtract 3 miles
    getAOI("Colorado Springs") %>% modify(-3)

## End(Not run)

[Package AOI version 0.1.9000 Index]