aoi_describe {AOI}	R Documentation

Describe an AOI

Description

Describe a spatial (sf/sp/raster) object in terms of a reproducable AOI (e.g. aoi_get) parameters.

Usage

aoi_describe(AOI, full = FALSE, km = FALSE)

Arguments

AOI
a spatial object (raster, sf, sp).

full
if TRUE, reverse geocoding descriptions returned, else just lat, lon, width, height, and origin (default = FALSE)

km
if TRUE, units are in kilometers, else in miles (default = FALSE)

Value

a data.frame of AOI descriptors including (at minimum):

lat

the AOI center latitude

lon

the AOI center longitude

height

height in (miles)

width

width in(miles)

origin

AOI origin

Examples

{
fname <- system.file("shape/nc.shp", package="sf")
nc <- sf::read_sf(fname)
aoi_describe(nc[1,])
}

[Package AOI version 0.1.9000 Index]