aoi_inside {AOI}	R Documentation

Is Inside

Description

A check to see if one object is inside another

Usage

aoi_inside(AOI, obj, total = TRUE)

Arguments

AOI
object 2

obj
object 1

total
boolean. If TRUE then check if obj is competely inside the AOI (and vice versa: order doesn't matter) If FALSE, then check if at least part of obj is in the AOI.

Value

boolean value

[Package AOI version 0.1.9000 Index]