geocodeOSM {AOI}	R Documentation

geocodeOSM

Description

Geocode via Open Street Maps API. geocodeOSM takes an input string and converts it to a geolocation. Addtionally it can return the location as a simple features point and the minimun bounding area of the location extent.

Usage

geocodeOSM(location, pt = FALSE, bb = FALSE, all = FALSE,
  full = FALSE)

Arguments

location
a place name

pt
if TRUE a simple feature point is appended to returned list

bb
if TRUE the OSM bounding area of the location is appended to returned list

all
if TRUE the point and bounding box representations are returned

full
logical. If TRUE all OSM attributes reuturned with query, else just the lat/long pair.

Value

at minimum a data.frame of lat, long

Author(s)

Mike Johnson

Examples

## Not run: 
geocodeOSM("UCSB")
geocodeOSM("Garden of the Gods", bb = TRUE)

## End(Not run)

[Package AOI version 0.1.9000 Index]