geocode_rev {AOI}	R Documentation

Reverse Geocoding

Description

Describe a location using the ERSI and OSM reverse geocoding web-services. This service provides tradional reverse geocoding (lat/lon to placename) but can also be use to get more information about a place name.

Usage

geocode_rev(x)

Arguments

x
a point provided by numeric lat/lon pair or character place name

Value

a data.frame of descriptive features

Author(s)

Mike Johnson

Examples

## Not run: 
 geocode_rev(x = c(38,-115)) %>% t()

 ```
 county          :	Lincoln County
 state           :	Nevada
 country         :	USA
 place_id        :	198776170
 osm_type        :	relation
 osm_id          :	166463
 lat             :	37.5449476
 lon             :	-114.8764448
 display_name    :	Lincoln County, Nevada, USA
 match_addr      :	89017, Hiko, Nevada
 longlabel       :	89017, Hiko, NV, USA
 shortlabel      :	89017
 addr_type       :	Postal
 city            :	Hiko
 lon             :	-115
 lat             :	38
 bb              :	-115.897545,-114.048473,36.8420756,38.678486
 ```

 geocode_rev("UCSB") %>% t()

 ````
 library         :	UCSB Library
 pedestrian      :	Library Plaza
 county          :	Santa Barbara County
 state           :	California
 postcode        :	93106
 country         :	USA
 place_id        :	156341322
 osm_type        :	way
 osm_id          :	355809608
 lat             :	34.41399165
 lon             :	-119.845522700258
 display_name    :	UCSB Library, Library Plaza, Santa Barbara County, California, 93106, USA
 match_addr      :	93106, Santa Barbara, California
 longlabel       :	93106, Santa Barbara, CA, USA
 city            :	Santa Barbara
 lat             :	34.4145937
 bb              :	-119.8458708,-119.8450475,34.4128884,34.414646
 ````

## End(Not run)

[Package AOI version 0.1.9000 Index]