Get area of interest (AOI) boundary

Get a Spatial* representation of an AOI defined by:

a US state name(s)

a US state, county pair(s)

a user Spatial, sf or raster object or

a clipping unit (see details)

getAOI wraps getFiat and getClip into a single function.

getAOI(state = NULL, county = NULL, clip = NULL, sf = FALSE,
  km = FALSE, bb = FALSE)

Arguments

state
character. Full name or two character abbriviation. Not case senstive

county
character. County name(s). Requires state input. Not case senstive

clip
Spatial object, a Raster object, or a list (see details and getClip)

sf
logical. If TRUE object returned is of class sf, default is FALSE and returns class SpatialPolygons

km
logical. If TRUE distance are in kilometers, default is FALSE and with distances in miles

bb
logical. If TRUE then the bounding geometry of state/county is returned, default is FALSE and returns fiat geometries

Value

a Spatial* object projected to EPSG:4269.

Details

If clip is a list, a clip unit requires a minimum of 3 inputs:

A point:

'location name' ex: "UCSB"

lat/lon pair: ex: '-36, -120'

A bounding box height

in miles ex: 10

A bounding box width

in miles ex: 10

The bounding box is always drawn in relation to the location. By default the point is treated as the center of the box. To define the realtive location of the point to the bounding box a fourth input can be used:

Origin

'center' (default)

'upperleft'

'upperright'

'lowerleft'

'lowerright'

3 to 5 elements can be used to paramaterize the clip element but ORDER MATTERS (point, height, width, origin). Acceptable variations include:

3 members: (1) location name, (2) height, (3) width

list("UCSB", 10, 10)

4 members: (1) latitude, (2) longitude, (3) height, (4) width

list(36, -120, 10, 10)

4 members: (1) location name, (2) height, (3) width, (4) origin

list("UCSB", 10, 10, "lowerright)

5 members: (1) lat, (2) long, (3) height, (4) width, (5) origin

list(36,-120, 10, 10, "upperright)

See also

Examples

# Get AOI defined by a state(s)
    getAOI(state = 'CA')

#> AOI defined as boundary of California

#> class       : SpatialPolygonsDataFrame 
#> features    : 1 
#> extent      : -124.4096, -114.1391, 32.53416, 42.00925  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0+no_defs 
#> variables   : 5
#> names       : statefp,  statens,    affgeoid, state_name, state_abbr 
#> value       :      06, 01779778, 0400000US06, California,         CA 

    getAOI(state = c('CA', 'nevada'))

#> AOI defined as boundary of California, and Nevada

#> class       : SpatialPolygonsDataFrame 
#> features    : 2 
#> extent      : -124.4096, -114.0399, 32.53416, 42.00925  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0+no_defs 
#> variables   : 5
#> names       : statefp,  statens,    affgeoid, state_name, state_abbr 
#> min values  :      06, 01779778, 0400000US06, California,         CA 
#> max values  :      32, 01779793, 0400000US32,     Nevada,         NV 

# Get AOI defined by state & county pair(s)
    getAOI(state = 'California', county = 'Santa Barbara')

#> AOI defined as boundary of Santa Barbara county California

#> class       : SpatialPolygonsDataFrame 
#> features    : 1 
#> extent      : -120.6708, -119.4423, 33.89571, 35.11468  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0+no_defs 
#> variables   : 7
#> names       : statefp, countyfp,       affgeoid, geoid,          name, state_name, state_abbr 
#> value       :      06,      083, 0500000US06083, 06083, Santa Barbara, California,         CA 

    getAOI(state = 'CA', county = c('Santa Barbara', 'ventura'))

#> AOI defined as boundary of Santa Barbara, and Ventura county California

#> class       : SpatialPolygonsDataFrame 
#> features    : 2 
#> extent      : -120.6708, -118.6335, 33.21543, 35.11468  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0+no_defs 
#> variables   : 7
#> names       : statefp, countyfp,       affgeoid, geoid,          name, state_name, state_abbr 
#> min values  :      06,      083, 0500000US06083, 06083, Santa Barbara, California,         CA 
#> max values  :      06,      111, 0500000US06111, 06111,       Ventura, California,         CA 

# Get AOI defined by external spatial file:
    getAOI(clip = rgdal::readOGR('la_metro.shp'))

#> Error in ogrListLayers(dsn = dsn): Cannot open data source

    getAOI(clip = raster('AOI.tif'))

#> Error in raster("AOI.tif"): could not find function "raster"

# Get AOI defined by 10 mile bounding box using lat/lon
    getAOI(clip = c(35, -119, 10, 10))

#> AOI defined as a 10 mile tall by 10 mile wide region centered on 35/-119 (lat/lon)

#> class       : SpatialPolygons 
#> features    : 1 
#> extent      : -119.0885, -118.9115, 34.92754, 35.07246  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0+no_defs 

# Get AOI defined by 10 mile2 bounding box using the 'KMART near UCSB' as lower left corner
    getAOI(clip = list('KMART near UCSB', 10, 10, 'lowerleft'))

#> AOI defined as a 10 mile tall by 10 mile wide region with (the) KMART near UCSB in the lower left corner

#> class       : SpatialPolygons 
#> features    : 1 
#> extent      : -119.8677, -119.692, 34.42859, 34.57352  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0+no_defs