getClip {AOI}	R Documentation

Convert clip unit ot geometry

Description

getClip generates a Spatial object based on a point; bounding box dimisions; and their relation to the point.

Usage

getClip(x, km = FALSE)

Arguments

location
Defined by a location or lat, long pair

height
define the height of the desired bounding box in miles

width
define the width of the desired bounding box in miles

origin
define the position of the point with respect to the bounding box. Default is set to center. Options include

"center"

"lowerleft"

"lowerright"

"upperright"

"upperleft"

Value

a SpatialPolygons object projected to EPSG:4269.

Author(s)

Mike Johnson

[Package AOI version 0.1.9000 Index]