AOI Package

An area of interest (AOI) is a geographic extent. The aim of this package is to help users create these - essentailly turning locations, place names, and political boundaries into servicable geometries for spatial analysis. The package is written using the simple features paradigm, projected to EPSG:4269.

The primary functions in this package are are geocode, geocode_rev, aoi_get, and bbox_get. The first returns a data.frame of corrdinates from place names using the OSM API; the second returns a list of descriptive features from a known place name or lat/lon pair; the third returns a spatial (sf) geometry from a country, state, county, or defined region, and the last an extent encompassing a set of input features. aoi_map which helps users visualize AOIs in a interactive leaflet map; and aoi_buffer which allows AOIs to be midified by uniform distances. Finally, aoi_describe breaks existing spatial features into aoi_get parameters to improve the reproducibility of geometry generation.

See the README on github, and the project webpage for examples here.