AOI Package

An area of interest (AOI) is a geographic extent and the aim of this package is to help users create these. The package is written using the simple features paradigm, however, by default, objects are returned as SpatialPolygons projected to EPSG:4269. For those that have made the jump to sf, all functions include a 'sf' parameter that can be set to TRUE and eventully the default behavior will change.

The primary functions to be aware of are geocode, getAOI, getBoundingBox The first returns a set spatial points, the second a single spatial geometry, and the last a geometry encompassing all input features. bbox_st and bbox_sp help convert AOIs between string and geometry manifestations; check helps users visualize AOIs in a interactive leaflet map; and buffer allows for the modification of AOIs by uniform distances. Finally, revgeocode provides a reverse geocoding interface to help understand coordinates as locations, and describe breaks existing spatial features into getAOI parameters to improve the reproducibility of geometry generation.

Two core datasets are served with the package reflecting the spatial geometries of US states and counties

See the README on github, and a webpage of examples here