Get Area of Interest (AOI) geometry

Generate a spatial geometry from:

Country name, 2-digit or 3-digit ISO abbreviation(s)

Country Region or Continent (e.g. South Asia, Africa)

US state name(s) or abbreviation(s)

US region (e.g. Northeast, South, North Central, West)

US state, county pair(s)

a spatial, sf or raster object

a clip unit (see details)

aoi_get(
  x = NULL,
  country = NULL,
  state = NULL,
  county = NULL,
  fip = NULL,
  km = FALSE,
  union = FALSE
)

Arguments

x

spatial, raster, sf or a list object (see details for list parameters)

country

character. Full name, ISO 3166-1 2 or 3 digit code. Not case senstive. Data comes from Natural Earth.

state

character. Full name or two character abbreviation. Not case sensitive. If state = 'conus', the lower 48 states will be returned. If state = 'all', all states will be returned.

county

character. County name(s). Requires state input. Not case sensitive If 'all' then all counties in a state are returned

fip

a 2 or 3 digit US fip code

km

logical. If TRUE distances are in kilometers, default is FALSE with distances in miles

union

logical. If TRUE objects are unioned into a single object

Value

a sf geometry projected to EPSG:4269.

Details

A x unit can be described by just a place name (e.g. 'UCSB'). In doing so the associated boundaries determined by geocode will be returned. To have greater control over the clip unit it can be defined as a list with a minimum of 3 inputs:

A point:

'place name' (character) ex: "UCSB" - or -

'lat/lon' pair: ex: "-36, -120"

A bounding box height (numeric)

in miles ex: 10

A bounding box width (numeric)

in miles ex: 10

The bounding box is always drawn in relation to the point. By default, the point is treated as the center of the box. To define the relative location of the point to the bounding box, a fourth input can be used:

Origin

'center' (default)

'upperleft'

'upperright'

'lowerleft'

'lowerright'

In total, 1 to 5 elements can be used to define clip element and ORDER MATTERS (point, height, width, origin). Acceptable variations include:

1 member: (1) place name

"UCSB"

1 member: (1) lat/lon pair

c(36, -119)

3 members: (1) location name, (2) height, (3) width

list("UCSB", 10, 10)

4 members: (1) lat, (2) lon, (3) height, (4) width

list(36, -120, 10, 10)

4 members: (1) place name, (2) height, (3) width, (4) origin

list("UCSB", 10, 10, "lowerright)

5 members: (1) lat, (2) lon, (3) height, (4) width, (5) origin

list(36,-120, 10, 10, "upperright)

Examples

if (FALSE) {
# Get AOI for a country
aoi_get(country = "Brazil")

# Get AOI for a location
aoi_get(x = "Sacramento")

# Get AOI defined by a state(s)
aoi_get(state = "CA")
aoi_get(state = c("CA", "nevada"))

# Get AOI defined by all states, or the lower 48
aoi_get(state = "all")
aoi_get(state = "conus")

# Get AOI defined by state & county pair(s)
aoi_get(state = "California", county = "Santa Barbara")
aoi_get(state = "CA", county = c("Santa Barbara", "ventura"))

# Get AOI defined by state & all counties
aoi_get(state = "California", county = "all")

# Get AOI defined by 10 mile bounding box using lat/lon
aoi_get(c(35, -119, 10, 10))

# Get AOI defined by 10 mile2 bounding box using the 'KMART near UCSB' as lower left corner
aoi_get(x = list("UCSB", 10, 10, "lowerleft"))
}