Is Inside

A check to see if one object is inside another

aoi_inside(AOI, obj, total = T)

Arguments

AOI
object 2

obj
object 1

total
boolean. If TRUE then check if obj is competely inside the AOI. If FALSE, then check if at least part of obj is in the AOI.

Value

boolean value

Examples

if (FALSE) {
  AOI = getAOI(state = "CA")
  obj = getAOI(state = "CA", county = "Santa Barbara")
  is_inside(AOI, obj)

  AOI = getAOI(state = "CA")
  obj = getAOI(state = "CO", county = "El Paso")
  is_inside(AOI, obj)
}