Geocoding

A wrapper around the OpenSteetMap geocoding web-services. Users can request a lat/lon pair, spatial points, and/or a bounding box geometries. One or more place name can be given at a time. If a single point is requested, `geocode` will provide a data.frame of lat/lon values and, if requested, a spatial point object and the geocode derived bounding box. If multiple place names are given, the returned objects will be a data.frame with columns for input name-lat-lon; if requested, a SpatialPoints object will be returned; and a minimum bounding box of all place names.

geocode(location = NULL, zipcode = NULL, event = NULL, pt = FALSE,
  bb = FALSE, all = FALSE, full = FALSE)

Arguments

location
character. Place name(s)

zipcode
character. USA zipcode(s)

event
character. a term to search for on wikipeida

pt
logical. If TRUE point geometery is appended to the returned list()

bb
logical. If TRUE bounding box geometry is appended to the returned list()

all
if TRUE the point and bounding box representations are returned

full
logical. If TRUE all OSM attributes reuturned with query, else just the lat/long pair.

Value

at minimum a data.frame of lat/lon coordinates. Possible list with appended spatial features of type sf or sp

Examples

if (FALSE) {
 ## geocode a single location
     geocode("UCSB")

 ## geocode a single location and return a SpatialPoints object
     geocode("UCSB", pt = TRUE)

 ## geocode a single location and derived bounding box of location
     geocode("UCSB", bb = TRUE)

 ## geocode multiple locations
     geocode(c("UCSB", "Goleta", "Santa Barbara"))

 ## geocode multiple points and generate a minimum bounding box of all locations and spatial points
     geocode(c("UCSB", "Goleta", "Santa Barbara"), bb = T, pt= T)
}