Get a bouding box for and AOI

getClip generates a Spatial object based on a point; bounding box dimisions; and their relation to the point.

getClip(location = NULL, width = NULL, height = NULL, origin = NULL)

Arguments

location
Defined by a location or lat, long pair

width
define the width of the desired bounding box in miles

height
define the height of the desired bounding box in miles

origin
define the position of the point with respect to the bounding box. Default is set to center. Options include

"center"

"lowerleft"

"lowerright"

"upperright"

"upperleft"

Value

a SpatialPolygons object projected to EPSG:4269.

Examples

# Get AOI defined by 10 mile bounding box using "UCSB" as the point
    getClip(location = "UCSB", width = 10, height = 10, origin = "center")

#> class       : SpatialPolygons 
#> features    : 1 
#> extent      : -119.9368, -119.7611, 34.3415, 34.48643  (xmin, xmax, ymin, ymax)
#> coord. ref. : +proj=longlat +ellps=GRS80 +towgs84=0,0,0,0,0,0,0+no_defs 

# Get AOI defined by 10 mile2 bounding box using the 'KMART near UCSB' as lower left corner
    getClip(clocation = NULL, width = NULL, height = NULL, origin = NULL)

#> Error in getClip(clocation = NULL, width = NULL, height = NULL, origin = NULL): unused argument (clocation = NULL)